Added default response
Previously, if user did not use WithResponseAs to specify the
unmarshal target, there will be a nil pointer error when run.
But some test may only concern the status code. It is non-sense
for user to specify response target for those tests.

Created a nilResp to cater the situation.

nilResp type will be default if no unmarshal target is set.
Any access of nilResp through Response interface will result in
error. Which reminds users to use WithResponseAs if their test(s)
need(s) them to.